C51-390: Refactor gulp file and gulp tasks
