+++ gulpfile.js
@@ -4,109 +4,53 @@
  * importing Shoreditch Partials and minifying the css file to .min.css
  *
  * Tasks
- * default : Runs SASS task
+ * default : Runs sass and test task
+ * sass:sync : Compiles civicase.scss under scss folder to CSS counterpart
  * sass: Compiles civicase.scss under scss folder to CSS counterpart
- * watch: Watches for scss file changes and run sass task
+ * test: Runs Karma unit tests
+ * backstopjs:reference: For creating reference screenshots
+ * backstopjs:test: For creating test screenshots and matching them
+ * backstopjs:openReport: For opening reports in the browser
+ * backstopjs:approve: Approving reports
+ * watch: Watches for scss and js file changes and run sass task
  */
+
 'use strict';
 
-var autoprefixer = require('gulp-autoprefixer');
-var bulk = require('gulp-sass-bulk-import');
-var civicrmScssRoot = require('civicrm-scssroot')();
-var cssmin = require('gulp-cssmin');
 var gulp = require('gulp');
-var karma = require('karma');
-var path = require('path');
-var postcss = require('gulp-postcss');
-var postcssDiscardDuplicates = require('postcss-discard-duplicates');
-var postcssPrefix = require('postcss-prefix-selector');
-var rename = require('gulp-rename');
-var sass = require('gulp-sass');
-var stripCssComments = require('gulp-strip-css-comments');
-var sourcemaps = require('gulp-sourcemaps');
-var transformSelectors = require('gulp-transform-selectors');
+
 var backstopJSTask = require('./gulp-tasks/backstopjs.js');
-
-var bootstrapNamespace = '#bootstrap-theme';
-var outsideNamespaceRegExp = /^\.___outside-namespace/;
-
-/**
-  * The gulp task updates and sync the scssRoot paths
-  */
-gulp.task('sass:sync', () => {
-  civicrmScssRoot.updateSync();
-});
+var sassTask = require('./gulp-tasks/sass.js');
+var sassSyncTask = require('./gulp-tasks/sass-sync.js');
+var testTask = require('./gulp-tasks/test.js');
+var watchTask = require('./gulp-tasks/watch.js');
 
 /**
- * The gulp task compiles and minifies scss/civicase.scss file into css/civicase.min.css.
- * Also prefix the output css selector with `#bootstrap-theme` selector except the output.
- * selector starts from either `body`, `page-civicrm-case` or `.___outside-namespace` classes.
+ * The gulp task updates and sync the scssRoot paths
  */
-gulp.task('sass', ['sass:sync'], function () {
-  return gulp.src('scss/civicase.scss')
-    .pipe(bulk())
-    .pipe(sourcemaps.init())
-    .pipe(autoprefixer({
-      browsers: ['last 2 versions'],
-      cascade: false
-    }))
-    .pipe(sass({
-      outputStyle: 'compressed',
-      includePaths: civicrmScssRoot.getPath(),
-      precision: 10
-    }).on('error', sass.logError))
-    .pipe(stripCssComments({ preserve: false }))
-    .pipe(postcss([postcssPrefix({
-      prefix: bootstrapNamespace + ' ',
-      exclude: [/^body/, /page-civicrm-case/, outsideNamespaceRegExp]
-    }), postcssDiscardDuplicates]))
-    .pipe(transformSelectors(removeOutsideNamespaceMarker, { splitOnCommas: true }))
-    .pipe(cssmin({ sourceMap: true }))
-    .pipe(rename({ suffix: '.min' }))
-    .pipe(sourcemaps.write('.'))
-    .pipe(gulp.dest('css/'));
-});
+gulp.task('sass:sync', sassSyncTask);
 
 /**
- * Watch task
+ * Gulp sass task
  */
-gulp.task('watch', function () {
-  gulp.watch('scss/**/*.scss', ['sass']);
-  gulp.watch(['ang/**/*.js', '!ang/test/karma.conf.js'], ['test']);
-  gulp.watch(civicrmScssRoot.getWatchList(), ['sass']);
-});
+gulp.task('sass', ['sass:sync'], sassTask);
 
 /**
- * Default task
+ * Gulp unit tests task
+ */
+gulp.task('test', testTask);
+
+/**
+ * Gulp watch task
+ */
+gulp.task('watch', watchTask);
+
+/**
+ * Gulp backstop tasks
+ */
+['reference', 'test', 'openReport', 'approve'].map(backstopJSTask);
+
+/**
+ * Gulp default task
  */
 gulp.task('default', ['sass', 'test']);
-
-/**
- * Deletes the special class that was used as marker for styles that should
- * not be nested inside the bootstrap namespace from the given selector
- *
- * @param  {String} selector
- * @return {String}
- */
-function removeOutsideNamespaceMarker (selector) {
-  return selector.replace(outsideNamespaceRegExp, '');
-}
-
-/**
- * Runs the unit tests
- */
-gulp.task('test', function (done) {
-  new karma.Server({
-    configFile: path.resolve(__dirname, 'ang/test/karma.conf.js'),
-    singleRun: true
-  }, done).start();
-});
-
-/**
-  * Gulp backstop tasks
-  * 'backstopjs:reference': For creating reference screenshots
-  * 'backstopjs:test': For creating test screenshots and matching them
-  * 'backstopjs:openReport': For opening reports in the browser
-  * 'backstopjs:approve': Approving reports
-  */
-['reference', 'test', 'openReport', 'approve'].map(backstopJSTask);
